requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
azure-functions==1.21.3
click==8.1.7; python_version >= '3.7'
fastapi==0.115.4; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.36; python_version >= '3.7'
starlette==0.41.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.32.0; python_version >= '3.8'
